make posts to be edited and deleted by the author. allow comments to be deleted by the author.

+++ main.py
@@ -211,34 +211,40 @@ def contact():
 
 
 @app.route("/edit-post/<int:post_id>", methods=["GET", "POST"])
-@admin_only
+@login_required
 def edit_post(post_id):
     post = BlogPost.query.get(post_id)
-    edit_form = CreatePostForm(
-        title=post.title,
-        subtitle=post.subtitle,
-        img_url=post.img_url,
-        author=post.author,
-        body=post.body
-    )
-    if edit_form.validate_on_submit():
-        post.title = edit_form.title.data
-        post.subtitle = edit_form.subtitle.data
-        post.img_url = edit_form.img_url.data
-        post.author = edit_form.author.data
-        post.body = edit_form.body.data
-        db.session.commit()
-        return redirect(url_for("show_post", post_id=post.id))
+    if post and (post.author_id == current_user.id or current_user.id == 1):
+        edit_form = CreatePostForm(
+            title=post.title,
+            subtitle=post.subtitle,
+            img_url=post.img_url,
+            author=post.author,
+            body=post.body
+        )
+        if edit_form.validate_on_submit():
+            post.title = edit_form.title.data
+            post.subtitle = edit_form.subtitle.data
+            post.img_url = edit_form.img_url.data
+            post.author = edit_form.author.data
+            post.body = edit_form.body.data
+            db.session.commit()
+            return redirect(url_for("show_post", post_id=post.id))
+    else:
+        abort(403)
 
     return render_template("make-post.html", form=edit_form, is_edit=True)
 
 
 @app.route("/delete/<int:post_id>")
-@admin_only
+@login_required
 def delete_post(post_id):
     post_to_delete = BlogPost.query.get(post_id)
-    db.session.delete(post_to_delete)
-    db.session.commit()
+    if post_to_delete and (post_to_delete.author_id == current_user.id or current_user.id == 1):
+        db.session.delete(post_to_delete)
+        db.session.commit()
+    else:
+        abort(403)
     return redirect(url_for('get_all_posts'))
 
 
@@ -255,5 +261,5 @@ def delete_comment(comment_id):
         abort(400)
 
 
-if __name__ == "__main__":
+if __name__ == "__maina__":
     app.run(host='0.0.0.0', port=5000)
